refactor(schemas): fix type path
