+++ packages/schemas/alterations/1.0.0_beta.14-1665300135-sign-in-sign-up.ts
@@ -1,6 +1,6 @@
 import { sql } from 'slonik';
 
-import type { AlterationScript } from '../src/types/alteration.js';
+import type { AlterationScript } from '../lib/types/alteration.js';
 
 enum SignInMethodState {
   Primary = 'primary',
